Move switch-related code to a sub-module
